% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherlink-package.R
\docType{package}
\name{weatherlink}
\alias{weatherlink}
\alias{weatherlink-package}
\title{Query and Orchestrate the Davis WeatherLink API}
\description{
Davis Instruments provides a public API (\url{https://weatherlink.github.io/v2-api/})
for owners and users of their weather datalogger products. Tools are provided to query
and orchestrate the Davis WeatherLink API (V2)
}
\references{
\href{https://weatherlink.github.io/v2-api/api-reference}{API Endpoint};
\href{https://weatherlink.github.io/v2-api/data-structure-types}{Data structure types}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/weatherlink}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/weatherlink/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
