% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-conditions.R, R/historic-conditions.R
\name{wl_conditions}
\alias{wl_conditions}
\title{Get current conditions data for one station}
\usage{
wl_conditions(
  station_id,
  api_key = Sys.getenv("WEATHERLINK_API_KEY"),
  api_secret = Sys.getenv("WEATHERLINK_API_SECRET")
)

wl_conditions(
  station_id,
  api_key = Sys.getenv("WEATHERLINK_API_KEY"),
  api_secret = Sys.getenv("WEATHERLINK_API_SECRET")
)
}
\arguments{
\item{station_id}{(chr) Station id}

\item{api_key, api_secret}{WeatherLink API (V2) key/secret. Generate & retrieve them from
\href{https://www.weatherlink.com/account}{your account page}. Automagically retrieved from
\code{WEATHERLINK_API_KEY} and \code{WEATHERLINK_API_SECRET} environment variables.}
}
\value{
list ith a \code{generated_at} attribute

list ith a \code{generated_at} attribute
}
\description{
Returns current conditions data for a weather station selected by a station ID.

Returns historic conditions data for a weather station selected by a station ID.
}
\references{
\href{https://weatherlink.github.io/v2-api/api-reference}{API Endpoint};
\href{https://weatherlink.github.io/v2-api/data-structure-types}{Data structure types}

\href{https://weatherlink.github.io/v2-api/api-reference}{API Endpoint};
\href{https://weatherlink.github.io/v2-api/data-structure-types}{Data structure types}
}
