% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor-activity.R
\name{wl_sensor_activity}
\alias{wl_sensor_activity}
\title{Get latest reporting times for all sensors attached to all weather stations associated with your API Key}
\usage{
wl_sensor_activity(
  api_key = Sys.getenv("WEATHERLINK_API_KEY"),
  api_secret = Sys.getenv("WEATHERLINK_API_SECRET")
)
}
\arguments{
\item{api_key, api_secret}{WeatherLink API (V2) key/secret. Generate & retrieve them from
\href{https://www.weatherlink.com/account}{your account page}. Automagically retrieved from
\code{WEATHERLINK_API_KEY} and \code{WEATHERLINK_API_SECRET} environment variables.}
}
\value{
data frame with a \code{generated_at} attribute
}
\description{
Returns the most recent times data was uploaded for all sensors associated with the api key that was passed. T
}
\references{
\href{https://weatherlink.github.io/v2-api/api-reference}{API Endpoint};
\href{https://weatherlink.github.io/v2-api/data-structure-types}{Data structure types}
}
